adicionado autorização no modelo crud

+++ src/actions/_model.ts
@@ -1,3 +1,5 @@
+import { ErrorsMessages } from "@/config/messages";
+import { ROLES } from "@/enums/roles";
 import { prisma } from "@/lib/prisma";
 import { DefaultReturn, ErrorReturn } from "@/types/actions/_general";
 import {
@@ -6,6 +8,7 @@ import {
   ListReturn,
 } from "@/types/actions/users";
 import { User as Model } from "@prisma/client";
+import { getServerSession } from "next-auth";
 
 const model = prisma.user
 const singular = 'usuário'
@@ -13,6 +16,12 @@ const plural = 'usuários'
 
 export async function create(data: CreateResource): Promise<DefaultReturn<Model> | ErrorReturn> {
   try {
+    const session = await getServerSession()
+
+    if (session?.user.role === ROLES.VIEWER) {
+      return { error: ErrorsMessages.not_authorized }
+    }
+
     const response = await model.create({ data })
     return { data: response }
   } catch (error) {
@@ -33,6 +42,12 @@ export async function find(): Promise<ListReturn | ErrorReturn> {
 
 export async function update(data: UpdateResource): Promise<DefaultReturn<Model> | ErrorReturn> {
   try {
+    const session = await getServerSession()
+
+    if (session?.user.role === ROLES.VIEWER) {
+      return { error: ErrorsMessages.not_authorized }
+    }
+
     const { id, ...dataToUpdate } = data
 
     const response = await model.update({ where: { id }, data: dataToUpdate })
@@ -45,6 +60,12 @@ export async function update(data: UpdateResource): Promise<DefaultReturn<Model>
 
 export async function remove(id: Model['id']): Promise<void | ErrorReturn> {
   try {
+    const session = await getServerSession()
+
+    if (session?.user.role === ROLES.VIEWER) {
+      return { error: ErrorsMessages.not_authorized }
+    }
+
     await model.delete({ where: { id } })
   } catch (error) {
     console.error(error)
